requirements.txt
Flask==2.0.1
google-generativeai==0.0.2
python-dotenv==0.19.2
gunicorn